build: template 모듈 분리(api/impl) 및 적용
- api: 템플릿 화면으로의 이동을 위한 TemplateNavKey 포함
- impl: 실제 템플릿 화면 구현 코드
