+++ android/Bottari/build.gradle.kts
@@ -10,6 +10,7 @@ plugins {
     alias(libs.plugins.kotlin.compose) apply false
     alias(libs.plugins.ksp) apply false
     alias(libs.plugins.hilt.android) apply false
+    alias(libs.plugins.serialization) apply false
 }
 
 allprojects {
