+++ android/Bottari/feature/template/api/.gitignore
