+++ android/Bottari/feature/template/api/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    alias(libs.plugins.bottari.android.feature.api)
+}
+
+android {
+    namespace = "com.bottari.feature.template.api"
+}
