+++ android/Bottari/feature/template/api/src/main/java/com/bottari/feature/template/api/TemplateNavKey.kt
@@ -1,0 +1,7 @@
+package com.bottari.feature.template.api
+
+import androidx.navigation3.runtime.NavKey
+import kotlinx.serialization.Serializable
+
+@Serializable
+data object TemplateNavKey : NavKey
