+++ android/Bottari/feature/template/impl/.gitignore
@@ -1,0 +1,1 @@
+/build
