+++ android/Bottari/feature/template/impl/src/main/java/com/bottari/feature/template/impl/TemplateScreen.kt
@@ -1,0 +1,81 @@
+package com.bottari.feature.template.impl
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.pager.PagerState
+import androidx.compose.foundation.pager.rememberPagerState
+import androidx.compose.material3.SnackbarHostState
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import com.bottari.bottari.designsystem.theme.BottariTheme
+import com.bottari.core.ui.component.BottariTabBar
+import com.bottari.feature.template.impl.bookmark.BookmarkTemplateScreen
+import com.bottari.feature.template.impl.main.MainTemplateScreen
+import com.bottari.feature.template.impl.my.MyTemplateScreen
+
+@Composable
+fun TemplateScreen(
+    snackbarState: SnackbarHostState,
+    navigateToTemplateDetail: (templateId: Long, isMyTemplate: Boolean, isBookmark: Boolean) -> Unit,
+    navigateToTemplateCreate: () -> Unit,
+) {
+    TemplatePager(
+        pageTitles = listOf("전체 템플릿", "나의 템플릿", "북마크"),
+        modifier = Modifier.padding(top = BottariTheme.spacing.spaceMedium),
+    ) { page ->
+        when (page) {
+            0 -> {
+                MainTemplateScreen(
+                    snackbarState = snackbarState,
+                    onClickDetail = { id -> navigateToTemplateDetail(id, false, false) },
+                )
+            }
+
+            1 -> {
+                MyTemplateScreen(
+                    snackbarState = snackbarState,
+                    onClickDetail = { id -> navigateToTemplateDetail(id, true, false) },
+                    onClickCreate = navigateToTemplateCreate,
+                )
+            }
+
+            2 -> {
+                BookmarkTemplateScreen(
+                    snackbarHostState = snackbarState,
+                    navigateToDetail = { id -> navigateToTemplateDetail(id, false, true) },
+                )
+            }
+        }
+    }
+}
+
+@Composable
+private fun TemplatePager(
+    pageTitles: List<String>,
+    modifier: Modifier = Modifier,
+    screen: @Composable (Int) -> Unit,
+) {
+    val pagerState: PagerState = rememberPagerState(initialPage = 0) { pageTitles.size }
+
+    Column {
+        BottariTabBar(
+            pageTitles = pageTitles,
+            pagerState = pagerState,
+            modifier = modifier,
+            screen = screen,
+        )
+    }
+}
+
+@Preview
+@Composable
+private fun TemplateScreenPreview() {
+    BottariTheme {
+        TemplateScreen(
+            snackbarState = SnackbarHostState(),
+            navigateToTemplateDetail = { _, _, _ -> },
+            navigateToTemplateCreate = {},
+        )
+    }
+}
