+++ android/Bottari/feature/template/impl/src/main/java/com/bottari/feature/template/impl/bookmark/BookmarkTemplateEvent.kt
@@ -1,0 +1,7 @@
+package com.bottari.feature.template.impl.bookmark
+
+sealed interface BookmarkTemplateEvent {
+    data object FetchBookmarkTemplateFailure : BookmarkTemplateEvent
+
+    data object DeleteBookmarkTemplateFailure : BookmarkTemplateEvent
+}
