+++ android/Bottari/feature/template/impl/src/main/java/com/bottari/feature/template/impl/main/MainTemplateUiEvent.kt
@@ -1,0 +1,11 @@
+package com.bottari.feature.template.impl.main
+
+sealed interface MainTemplateUiEvent {
+    data object SearchTemplateSuccess : MainTemplateUiEvent
+
+    data object FetchBottariTemplatesFailure : MainTemplateUiEvent
+
+    data object AddBookmarkFailure : MainTemplateUiEvent
+
+    data object DeleteBookmarkFailure : MainTemplateUiEvent
+}
