+++ android/Bottari/feature/template/impl/src/main/java/com/bottari/feature/template/impl/my/MyTemplateUiEvent.kt
@@ -1,0 +1,7 @@
+package com.bottari.feature.template.impl.my
+
+sealed interface MyTemplateUiEvent {
+    data object FetchBottariTemplatesFailure : MyTemplateUiEvent
+
+    data object DeleteTemplateFailure : MyTemplateUiEvent
+}
