+++ android/Bottari/feature/template/impl/src/main/java/com/bottari/feature/template/impl/navigation/TemplateEntryBuilder.kt
@@ -1,0 +1,44 @@
+package com.bottari.feature.template.impl.navigation
+
+import androidx.navigation3.runtime.EntryProviderScope
+import androidx.navigation3.runtime.NavKey
+import com.bottari.core.navigation.LocalNavigator
+import com.bottari.core.ui.provider.LocalSnackbarHostState
+import com.bottari.feature.template.api.TemplateNavKey
+import com.bottari.feature.template.create.navigation.TemplateCreateNavKey
+import com.bottari.feature.template.detail.navigation.TemplateDetailNavKey
+import com.bottari.feature.template.impl.TemplateScreen
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.android.components.ActivityRetainedComponent
+import dagger.multibindings.IntoSet
+
+fun EntryProviderScope<NavKey>.templateEntryBuilder() {
+    entry<TemplateNavKey> {
+        val navigator = LocalNavigator.current
+
+        TemplateScreen(
+            snackbarState = LocalSnackbarHostState.current,
+            navigateToTemplateDetail = { id, isMyTemplate, isBookmark ->
+                TemplateDetailNavKey(
+                    templateId = id,
+                    isMyTemplate = isMyTemplate,
+                    isBookmark = isBookmark,
+                ).let(navigator::navigate)
+            },
+            navigateToTemplateCreate = { navigator.navigate(TemplateCreateNavKey) },
+        )
+    }
+}
+
+@Module
+@InstallIn(ActivityRetainedComponent::class)
+object TemplateEntryBuilderModule {
+    @IntoSet
+    @Provides
+    fun provideTemplateEntryBuilder(): EntryProviderScope<NavKey>.() -> Unit =
+        {
+            templateEntryBuilder()
+        }
+}
