+++ android/Bottari/feature/template/impl/src/main/res/values/strings.xml
