--- android/Bottari/feature/template/src/main/java/com/bottari/feature/template/bookmark/BookmarkTemplateEvent.kt
@@ -1,7 +1,0 @@
-package com.bottari.feature.template.bookmark
-
-sealed interface BookmarkTemplateEvent {
-    data object FetchBookmarkTemplateFailure : BookmarkTemplateEvent
-
-    data object DeleteBookmarkTemplateFailure : BookmarkTemplateEvent
-}
