--- android/Bottari/feature/template/src/main/java/com/bottari/feature/template/component/TemplateItem.kt
@@ -1,299 +1,0 @@
-package com.bottari.feature.template.component
-
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.FlowRow
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.wrapContentHeight
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.rounded.List
-import androidx.compose.material.icons.filled.Download
-import androidx.compose.material.icons.filled.Person
-import androidx.compose.material3.Icon
-import androidx.compose.material3.Text
-import androidx.compose.material3.ripple
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.bottari.bottari.designsystem.component.BottariCard
-import com.bottari.bottari.designsystem.theme.BottariTheme
-import com.bottari.core.ui.component.CollapsedListLine
-import com.bottari.core.ui.model.template.BottariTemplateHashtagUiModel
-
-@Composable
-fun TemplateItem(
-    title: String,
-    description: String,
-    items: List<String>,
-    author: String,
-    takenCount: Int,
-    hashtags: List<BottariTemplateHashtagUiModel>,
-    onClickHashtag: (BottariTemplateHashtagUiModel) -> Unit,
-    modifier: Modifier = Modifier,
-    iconButton: @Composable () -> Unit,
-) {
-    val hashTagSectionPadding =
-        if (hashtags.isNotEmpty()) BottariTheme.spacing.spaceMedium else 0.dp
-
-    BottariCard(
-        modifier = modifier,
-        contentPadding = PaddingValues(),
-    ) {
-        Column(modifier = Modifier.padding(BottariTheme.spacing.spaceMedium)) {
-            TemplateItemHeader(
-                title = title,
-                description = description,
-                items = items,
-                iconButton = iconButton,
-            )
-
-            if (author.isNotBlank() && takenCount != -1) {
-                Spacer(modifier = Modifier.height(BottariTheme.spacing.spaceSmall))
-                TemplateHashtagSection(hashtags = hashtags, onClickHashtag = onClickHashtag)
-                Spacer(modifier = Modifier.height(hashTagSectionPadding))
-                TemplateItemFooter(author = author, takenCount = takenCount, itemCount = items.size)
-            }
-        }
-    }
-}
-
-@Composable
-private fun TemplateItemHeader(
-    title: String,
-    description: String,
-    items: List<String>,
-    modifier: Modifier = Modifier,
-    iconButton: @Composable () -> Unit,
-) {
-    Row(
-        modifier =
-            modifier
-                .fillMaxWidth()
-                .wrapContentHeight(),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.Top,
-    ) {
-        Column(modifier = Modifier.weight(1f)) {
-            Text(
-                text = title,
-                style = BottariTheme.typography.semiBold20.toTextStyle(),
-                color = BottariTheme.colors.black,
-            )
-            Spacer(Modifier.height(BottariTheme.spacing.spaceXSmall))
-
-            if (description.isBlank()) {
-                CollapsedListLine(
-                    items = items,
-                    textStyle =
-                        BottariTheme.typography.medium14
-                            .toTextStyle()
-                            .copy(color = BottariTheme.colors.gray500),
-                )
-            } else {
-                Text(
-                    text = description,
-                    style = BottariTheme.typography.medium14.toTextStyle(),
-                    color = BottariTheme.colors.gray500,
-                )
-            }
-        }
-
-        Spacer(modifier = Modifier.width(BottariTheme.spacing.spaceSmall))
-
-        iconButton()
-    }
-}
-
-@Composable
-private fun TemplateHashtagSection(
-    hashtags: List<BottariTemplateHashtagUiModel>,
-    onClickHashtag: (BottariTemplateHashtagUiModel) -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    FlowRow(
-        modifier = modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.spacedBy(BottariTheme.spacing.space2xSmall),
-        verticalArrangement = Arrangement.spacedBy(BottariTheme.spacing.spaceXSmall),
-        itemVerticalAlignment = Alignment.CenterVertically,
-    ) { hashtags.forEach { hashtag -> HashtagChip(hashtag, onClickHashtag) } }
-}
-
-@Composable
-private fun HashtagChip(
-    hashtag: BottariTemplateHashtagUiModel,
-    onClickHashtag: (BottariTemplateHashtagUiModel) -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val chipShape = BottariTheme.shapes.pill
-
-    Box(
-        contentAlignment = Alignment.Center,
-        modifier =
-            modifier
-                .clip(chipShape)
-                .background(
-                    color = Color(0xFFEFF6FF),
-                    shape = chipShape,
-                ).clickable(
-                    interactionSource = remember { MutableInteractionSource() },
-                    indication = ripple(color = BottariTheme.colors.primary),
-                ) { onClickHashtag(hashtag) }
-                .padding(
-                    vertical = BottariTheme.spacing.space2xSmall,
-                    horizontal = BottariTheme.spacing.spaceXSmall,
-                ),
-    ) {
-        Text(
-            text = "#${hashtag.name.replace(" ", "")}",
-            style = BottariTheme.typography.medium12.toTextStyle(),
-            color = BottariTheme.colors.primary,
-        )
-    }
-}
-
-@Composable
-private fun TemplateItemFooter(
-    author: String,
-    takenCount: Int,
-    itemCount: Int,
-    modifier: Modifier = Modifier,
-) {
-    val contentColor = BottariTheme.colors.gray600
-
-    Row(
-        modifier =
-            modifier
-                .fillMaxWidth()
-                .height(20.dp),
-        horizontalArrangement = Arrangement.SpaceBetween,
-        verticalAlignment = Alignment.Bottom,
-    ) {
-        Row(verticalAlignment = Alignment.CenterVertically) {
-            Icon(
-                imageVector = Icons.Default.Person,
-                contentDescription = null,
-                tint = contentColor,
-            )
-            Spacer(Modifier.width(BottariTheme.spacing.space2xSmall))
-            Text(
-                text = author,
-                style = BottariTheme.typography.medium12.toTextStyle(),
-                color = contentColor,
-            )
-        }
-
-        Row(verticalAlignment = Alignment.CenterVertically) {
-            Icon(
-                imageVector = Icons.Default.Download,
-                contentDescription = null,
-                tint = contentColor,
-            )
-            Spacer(Modifier.width(BottariTheme.spacing.space2xSmall))
-            Text(
-                text = takenCount.toString(),
-                style = BottariTheme.typography.medium12.toTextStyle(),
-                color = contentColor,
-            )
-            Spacer(Modifier.width(BottariTheme.spacing.spaceSmall))
-            Icon(
-                imageVector = Icons.AutoMirrored.Rounded.List,
-                modifier = Modifier.size(height = 24.dp, width = 20.dp),
-                contentDescription = null,
-                tint = contentColor,
-            )
-            Spacer(Modifier.width(BottariTheme.spacing.space2xSmall))
-            Text(
-                text = itemCount.toString(),
-                style = BottariTheme.typography.medium12.toTextStyle(),
-                color = contentColor,
-            )
-        }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun TemplateItemPreview() {
-    BottariTheme {
-        Column(
-            modifier = Modifier.padding(12.dp),
-            verticalArrangement = Arrangement.spacedBy(8.dp),
-        ) {
-            TemplateItem(
-                title = "신입사원 온보딩 가이드",
-                description = "새로운 직장 생활을 위한 완벽 가이드",
-                author = "다이스",
-                takenCount = 100024,
-                items = List(10) { "아이템 $it" },
-                hashtags = List(3) { BottariTemplateHashtagUiModel(it.toLong(), "해시태그$it") },
-                onClickHashtag = {},
-            ) {
-                TemplateItemIconButton(
-                    type = TemplateItemType.MyTemplate,
-                    onClick = {},
-                )
-            }
-
-            TemplateItem(
-                title = "신입사원 온보딩 가이드",
-                description = "새로운 직장 생활을 위한 완벽 가이드",
-                author = "",
-                takenCount = 0,
-                items = List(10) { "아이템 $it" },
-                hashtags = List(3) { BottariTemplateHashtagUiModel(it.toLong(), "해시태그 $it") },
-                onClickHashtag = {},
-            ) {
-                TemplateItemIconButton(
-                    type = TemplateItemType.Bookmark(true),
-                    onClick = {},
-                )
-            }
-
-            TemplateItem(
-                title = "신입사원 온보딩 가이드",
-                description = "새로운 직장 생활을 위한 완벽 가이드",
-                author = "다이스",
-                takenCount = 100024,
-                items = List(10) { "아이템 $it" },
-                hashtags = List(3) { BottariTemplateHashtagUiModel(it.toLong(), "해시태그 $it") },
-                onClickHashtag = {},
-            ) {
-                TemplateItemIconButton(
-                    type = TemplateItemType.Bookmark(false),
-                    onClick = {},
-                )
-            }
-
-            TemplateItem(
-                title = "신입사원 온보딩 가이드",
-                description = "새로운 직장 생활을 위한 완벽 가이드",
-                author = "",
-                takenCount = 0,
-                items = List(10) { "아이템 $it" },
-                hashtags = emptyList(),
-                onClickHashtag = {},
-            ) {
-                TemplateItemIconButton(
-                    type = TemplateItemType.Bookmark(false),
-                    onClick = {},
-                )
-            }
-        }
-    }
-}
