--- android/Bottari/feature/template/src/main/java/com/bottari/feature/template/main/MainTemplateUiEvent.kt
@@ -1,11 +1,0 @@
-package com.bottari.feature.template.main
-
-sealed interface MainTemplateUiEvent {
-    data object SearchTemplateSuccess : MainTemplateUiEvent
-
-    data object FetchBottariTemplatesFailure : MainTemplateUiEvent
-
-    data object AddBookmarkFailure : MainTemplateUiEvent
-
-    data object DeleteBookmarkFailure : MainTemplateUiEvent
-}
