--- android/Bottari/feature/template/src/main/java/com/bottari/feature/template/my/MyTemplateUiEvent.kt
@@ -1,7 +1,0 @@
-package com.bottari.feature.template.my
-
-sealed interface MyTemplateUiEvent {
-    data object FetchBottariTemplatesFailure : MyTemplateUiEvent
-
-    data object DeleteTemplateFailure : MyTemplateUiEvent
-}
