--- android/Bottari/feature/template/src/main/java/com/bottari/feature/template/my/MyTemplateUiState.kt
@@ -1,15 +1,0 @@
-package com.bottari.feature.template.my
-
-import androidx.compose.runtime.Immutable
-import com.bottari.core.ui.model.template.BottariTemplateUiModel
-
-@Immutable
-data class MyTemplateUiState(
-    val isLoading: Boolean = false,
-    val isFetched: Boolean = false,
-    val isRefreshing: Boolean = false,
-    val templates: List<BottariTemplateUiModel> = emptyList(),
-) {
-    val isEmpty: Boolean = isFetched && templates.isEmpty()
-    val showLoading: Boolean = isLoading && isRefreshing.not()
-}
