--- android/Bottari/feature/template/src/main/java/com/bottari/feature/template/navigation/TemplateNavKey.kt
@@ -1,7 +1,0 @@
-package com.bottari.feature.template.navigation
-
-import com.bottari.core.navigation.MainTabNavKey
-import kotlinx.serialization.Serializable
-
-@Serializable
-data object TemplateNavKey : MainTabNavKey
