+++ android/Bottari/settings.gradle.kts
@@ -39,7 +39,8 @@ include(":core:navigation")
 include(":core:common")
 
 include(":feature:mybottari")
-include(":feature:template")
+include(":feature:template:api")
+include(":feature:template:impl")
 include(":feature:template-detail")
 include(":feature:template-create")
 include(":feature:personal-checklist")
